+++ src/templates/document/script.js
@@ -7,10 +7,7 @@ var Paris = window.Paris;
 Paris.document = (function(){
 
   var defaultOptions = {
-    postitContentUrl: '/postit/${pageId}',
-    isUserSubscribedUrl: '/alerts/isUserSubscribed',
-    subscribeUserUrl: '/alerts/subscribeUser',
-    unsubscribeUserUrl: '/alerts/unsubscribeUser'
+    postitContentUrl: '/postit/${pageId}'
   };
 
   function document(selector, userOptions){
@@ -27,7 +24,6 @@ Paris.document = (function(){
 
       if (Cookies.get(Paris.config.cookies.parisconnect.name)) {
         loadPostit();
-        setupAlertAjaxOperations();
       }
     }
 
@@ -58,54 +54,6 @@ Paris.document = (function(){
       });
     }
 
-    function setupAlertAjaxOperations() {
-      // init for alert
-      var $alertIcon = $('.document-heading-icons .icon-bell');
-      var alertId = $alertIcon.attr('href') && $alertIcon.attr('href').split('/').pop();
-
-      if (!alertId) {return;}
-
-      // get alert current status
-      var isSubscribed;
-      $.ajax({
-        type: 'get',
-        url: options.isUserSubscribedUrl,
-        data: {idalertes: alertId},
-        success: function (subscribed) {
-          switch (subscribed) {
-            case false:
-              isSubscribed = false;
-              break;
-            case true:
-              isSubscribed = true;
-              break;
-            default:
-              console.error('response to ' + options.isUserSubscribedUrl + '?idalertes=' + alertId + 'is not recognized: ' + subscribed);
-              return;
-          }
-          $alertIcon.attr('href', 'javascript:void(0)');
-          setupAlertHandlers();
-        }
-      });
-
-      function setupAlertHandlers() {
-        $alertIcon.attr('title', isSubscribed ? Paris.i18n.t("alerts/unsubscribe") : Paris.i18n.t("alerts/subscribe"));
-        $alertIcon.toggleClass('active', isSubscribed);
-        $alertIcon.on('click', function onClick() {
-          $.ajax({
-            type: 'get',
-            url: isSubscribed ? options.unsubscribeUserUrl :  options.subscribeUserUrl,
-            data: {idalertes: alertId},
-            success: function () {
-              isSubscribed = !isSubscribed;
-              $alertIcon.unbind('click');
-              setupAlertHandlers();
-            }
-          });
-        });
-      }
-
-    }
 
     init();